#84 Enhance Expenses module components
